app/requirements.txt
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
django_filter==24.3
djangorestframework-simplejwt==5.3.1
django-ckeditor==6.7.2
psycopg2-binary==2.9.9
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.43.0
Pillow==10.4.0
drf-spectacular==0.27.2
gunicorn==20.1.0
requests==2.32.3
django-unfold==0.45.0